refactor: extract decompose grouping planner
